메인뷰 높이조정및 바인딩 완성
메인뷰 높이조정및 필요한 바인딩들 추가해서
메인뷰 기능 자체는 완성이라고 봐도될듯합니다.
남은건 코드 클린작업이랑 혹시 기능필요하면 조금씩만 추가하면될것같습니다.

+++ app/src/main/java/com/uos/vcommcerce/datamodel/ProductDTO.kt
@@ -20,7 +20,7 @@ data class ProductDTO (
     //제품 설명
     var productExplain : String? = null,
     //리뷰 평점 총합
-    var totalRating : Long? = null,
+    var totalRating : Float? = null,
     //평점 갯수
     var ratingCount : Long? = null,
     //업로드 시간
